backend/pyproject.toml
[project]
name = "hirelense-ai-backend"
version = "0.1.0"
description = "Backend API for HireLens AI."
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.115.12",
    "uvicorn[standard]==0.34.2",
    "python-dotenv==1.0.1",
    "sqlalchemy==2.0.40",
    "alembic==1.15.2",
    "passlib==1.7.4",
    "PyJWT==2.10.1",
    "psycopg[binary]==3.2.6",
]

[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "httpx==0.28.1",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
